+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- [2025-02-14] [apiKeyAuthorizationProviders, vercel, brex](https://github.com/RubricLab/auth/commit/498233933eccab4cb147a0b4bd8a01f75ccc2a75)
 - [2025-02-13] [github expire tokens later, delete account if refresh fails](https://github.com/RubricLab/auth/commit/8a873404076151a7f0eaf490a8d8e9c483349d56)
 - [2025-02-13] [biome shiz](https://github.com/RubricLab/auth/commit/c9e3cc3f887614a11907fcfbc137092bed3cfcf8)
 - [2025-02-13] [github providers](https://github.com/RubricLab/auth/commit/e98d787af8ac54d174fddd8c5938b960d5d72462)